Break up aiohttp tests
